refactor: split production_order_service into focused modules (DEBT-1 D2-A)
Mechanical split of backend/app/services/production_order_service.py
(2,056 lines) per DEBT-1 D2-A. Move-only; no behavior change.

- production_order_release_service.py: release_production_order + release
  gating + self-heal helpers (_has_active_reservation,
  _maybe_backfill_op_material_allocation, _maybe_reserve_missing_materials).
- production_order_execution_service.py: start/complete/accept-short/split
  lifecycle transitions.
- production_order_service.py keeps the full public surface via a
  bottom-of-file re-export block so all existing callers (endpoints,
  sales_order_*, dispatch_service, inventory_service, tests) work unchanged.

Submodules import core helpers lazily inside function bodies (matching the
file's existing lazy cross-service import convention) to avoid a circular
import when a submodule is imported before core.

Deviation: record_scrap and scrap-reason CRUD stay in core because the test
suite patches "app.services.production_order_service.ScrapRecord" by string
path to intercept the constructor inside record_scrap; moving it would break
the patch and require rewriting the test (forbidden by the mechanical-split
Rules of Engagement).

ruff clean; full pytest: 5539 passed, 13 skipped, 0 failed.

Agent-Session: d2a-prodorder
