+++ backend/app/services/production_order_service.py
@@ -25,7 +25,6 @@ from app.models.inventory import Inventory
 from app.models.manufacturing import Routing, RoutingOperation, Resource
 from app.models.production_order import ProductionOrderOperationMaterial, ScrapRecord
 from app.models.work_center import WorkCenter
-from app.models.inventory import InventoryTransaction
 from app.models.material_spool import MaterialSpool, ProductionOrderSpool
 from app.services.supply_netting import get_projected_available
 logger = get_logger(__name__)
@@ -330,571 +329,6 @@ def delete_production_order(db: Session, order_id: int) -> None:
 # =============================================================================
 # Status Management
 # =============================================================================
-
-# ---------------------------------------------------------------------------
-# Private helpers for release_production_order
-# ---------------------------------------------------------------------------
-
-def _has_active_reservation(
-    db: Session,
-    production_order_id: int,
-    component_id: int,
-) -> bool:
-    """Return True if there is a net-positive ledger reservation for
-    (production_order_id, component_id), i.e. at least one 'reservation'
-    transaction exists and its quantity exceeds any 'reservation_release'
-    transactions for the same pair.
-    """
-    from sqlalchemy import func as sqlfunc
-    from sqlalchemy import case
-
-    net = (
-        db.query(
-            sqlfunc.coalesce(
-                sqlfunc.sum(
-                    case(
-                        (InventoryTransaction.transaction_type == "reservation",
-                         InventoryTransaction.quantity),
-                        else_=Decimal("0"),
-                    )
-                    - case(
-                        (InventoryTransaction.transaction_type == "reservation_release",
-                         InventoryTransaction.quantity),
-                        else_=Decimal("0"),
-                    )
-                ),
-                Decimal("0"),
-            )
-        )
-        .filter(
-            InventoryTransaction.reference_type == "production_order",
-            InventoryTransaction.reference_id == production_order_id,
-            InventoryTransaction.product_id == component_id,
-            InventoryTransaction.transaction_type.in_(
-                ["reservation", "reservation_release"]
-            ),
-        )
-        .scalar()
-    )
-    return (net or Decimal("0")) > Decimal("0")
-
-
-def _maybe_backfill_op_material_allocation(
-    db: Session,
-    order: "ProductionOrder",
-) -> None:
-    """Self-heal: if active ledger reservations exist for *order* but all
-    op-material rows have quantity_allocated=0, rebuild from the ledger.
-
-    This repairs the 17 draft WOs created before fix #715 the moment an
-    operator tries to release them — no data migration needed.  The heal
-    is idempotent: healthy orders (rows already populated) are untouched.
-    """
-    has_op_mats = any(
-        mat
-        for op in order.operations
-        for mat in op.materials
-    )
-    if not has_op_mats:
-        return
-
-    all_zero = all(
-        Decimal(str(mat.quantity_allocated)) == Decimal("0")
-        for op in order.operations
-        for mat in op.materials
-    )
-    if not all_zero:
-        # Already populated — nothing to do.
-        return
-
-    # Check if any active reservation ledger rows exist for this order.
-    has_reservations = (
-        db.query(InventoryTransaction.id)
-        .filter(
-            InventoryTransaction.reference_type == "production_order",
-            InventoryTransaction.reference_id == order.id,
-            InventoryTransaction.transaction_type == "reservation",
-        )
-        .first()
-    ) is not None
-
-    if not has_reservations:
-        return
-
-    # Import here to avoid circular dependency at module level.
-    from app.services.inventory_service import _backfill_op_material_from_ledger
-    _backfill_op_material_from_ledger(db, order)
-    # Flush so the gate check below (and any subsequent db.refresh) sees the
-    # updated rows immediately — the surrounding call-site holds the transaction.
-    db.flush()
-
-
-def _maybe_reserve_missing_materials(
-    db: Session,
-    order: "ProductionOrder",
-    user_email: str,
-) -> None:
-    """RESERVE-1 level-2 self-heal: reserve at release time when reservation
-    never ran or under-ran.
-
-    Level 1 (_maybe_backfill_op_material_allocation) only repairs ledger↔row
-    desync — it cannot help when the ledger itself has no (or insufficient)
-    reservations.  That happens for brownfield orders whose creation-time
-    reservation walked an empty/production-line-less BOM while the materials
-    actually live on the routing (op-material rows).
-
-    If any op-material row still shows quantity_allocated < quantity_required
-    after level 1, the net ledger reservation for that component is missing
-    or short, so run reserve_production_materials — it is delta-safe
-    (RESERVE-1) and tops up only the shortfall.  HARD-5 semantics are
-    preserved: zero-stock components reserve ahead of receipt (flag, not
-    block) and release proceeds.
-    """
-    rows = [mat for op in order.operations for mat in op.materials]
-    if not rows:
-        return
-
-    shortfall_rows = [
-        mat for mat in rows
-        if Decimal(str(mat.quantity_allocated or 0))
-        < Decimal(str(mat.quantity_required or 0))
-    ]
-    if not shortfall_rows:
-        return
-
-    never_ran = all(
-        Decimal(str(mat.quantity_allocated or 0)) == Decimal("0")
-        for mat in rows
-    )
-    if never_ran:
-        logger.info(
-            "RESERVE-1 self-heal: reservation had never run for PO#%s — "
-            "reserving at release",
-            order.code,
-        )
-    else:
-        logger.info(
-            "RESERVE-1 self-heal: reservation under-ran for PO#%s "
-            "(%d op-material row(s) short) — topping up at release",
-            order.code,
-            len(shortfall_rows),
-        )
-
-    from app.services.inventory_service import reserve_production_materials
-    reserve_production_materials(db, order, created_by=user_email)
-    # Flush so the gate re-check reads the updated rows.
-    db.flush()
-
-
-def release_production_order(
-    db: Session,
-    order_id: int,
-    user_email: str,
-    force: bool = False,
-) -> ProductionOrder:
-    """
-    Release a production order for manufacturing.
-
-    Validates material availability and transitions status to 'released'.
-    """
-    order = get_production_order(db, order_id)
-
-    # Idempotent: already released is a no-op
-    if order.status == "released":
-        return order
-
-    # Allow releasing from draft, scheduled, or on_hold (resume from hold)
-    if order.status not in ["draft", "scheduled", "on_hold"]:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Cannot release order in {order.status} status"
-        )
-
-    # Self-heal level 1: if the order has active ledger reservations but
-    # op-material rows still show quantity_allocated=0, the reservation sync
-    # was skipped (e.g. orders created before fix #715, or a future code path
-    # that calls reserve_production_materials without the sync step).
-    # Backfill now so the gate below reads truth rather than always-zero.
-    _maybe_backfill_op_material_allocation(db, order)
-
-    # Check material availability unless forced
-    if not force:
-        # Self-heal level 2 (RESERVE-1): if rows still show a shortfall after
-        # the ledger backfill, reservation never ran or under-ran (e.g. pure
-        # routing-material products whose creation-time reservation walked an
-        # empty BOM).  reserve_production_materials is delta-safe, so this
-        # tops up only the missing quantities, then the gate re-checks.
-        _maybe_reserve_missing_materials(db, order, user_email)
-
-        blocking_issues = []
-        for op in order.operations:
-            for mat in op.materials:
-                if mat.quantity_allocated < mat.quantity_required:
-                    component = db.query(Product).filter(
-                        Product.id == mat.component_id
-                    ).first()
-                    sku = component.sku if component else f"ID:{mat.component_id}"
-                    # Distinguish "never reserved" from "genuinely short"
-                    # by checking whether any active reservations exist for
-                    # this (order, component) pair.
-                    has_any_reservation = _has_active_reservation(
-                        db, order.id, mat.component_id
-                    )
-                    if has_any_reservation:
-                        blocking_issues.append({
-                            "component_sku": sku,
-                            "operation": op.operation_name,
-                            "needed": float(mat.quantity_required),
-                            "reserved": float(mat.quantity_allocated),
-                            "reason": "insufficient_stock",
-                        })
-                    else:
-                        blocking_issues.append({
-                            "component_sku": sku,
-                            "operation": op.operation_name,
-                            "needed": float(mat.quantity_required),
-                            "reserved": 0.0,
-                            "reason": "not_allocated",
-                        })
-
-        if blocking_issues:
-            not_allocated = [
-                b for b in blocking_issues if b["reason"] == "not_allocated"
-            ]
-            short = [
-                b for b in blocking_issues if b["reason"] == "insufficient_stock"
-            ]
-
-            if not_allocated and not short:
-                # Nearly unreachable after the level-2 self-heal above: it
-                # fires only when reservation itself errored for every
-                # shortfall component (e.g. UOMConversionError skip).
-                message = (
-                    "Material reservation failed during release — check "
-                    "component/UOM configuration"
-                )
-            else:
-                parts = []
-                for b in short:
-                    parts.append(
-                        f"Insufficient stock for {b['component_sku']}: "
-                        f"need {b['needed']}, reserved {b['reserved']}"
-                    )
-                message = "; ".join(parts) if parts else "Cannot release: material shortages detected"
-
-            raise HTTPException(
-                status_code=400,
-                detail={
-                    "message": message,
-                    "shortages": blocking_issues,
-                    "hint": "Use force=true to release anyway",
-                }
-            )
-
-    order.status = "released"
-    order.released_at = datetime.now(timezone.utc)
-
-    logger.info(f"Production order {order.code} released by {user_email}")
-
-    return order
-
-
-def start_production_order(db: Session, order_id: int) -> ProductionOrder:
-    """Start production on an order."""
-    order = get_production_order(db, order_id)
-
-    if order.status not in ["released", "scheduled"]:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Cannot start order in {order.status} status"
-        )
-
-    order.status = "in_progress"
-    order.actual_start = datetime.now(timezone.utc)
-
-    # Start first operation if not already started
-    first_op = (
-        db.query(ProductionOrderOperation)
-        .filter(ProductionOrderOperation.production_order_id == order_id)
-        .order_by(ProductionOrderOperation.sequence)
-        .first()
-    )
-    if first_op and first_op.status == "pending":
-        first_op.status = "running"
-        first_op.actual_start = datetime.now(timezone.utc)
-
-    return order
-
-
-def complete_production_order(
-    db: Session,
-    order_id: int,
-    user_email: str,
-    quantity_good: int,
-    quantity_scrapped: int = 0,
-    force_close_short: bool = False,
-    notes: Optional[str] = None,
-) -> ProductionOrder:
-    """
-    Complete a production order and record finished goods to inventory.
-
-    Args:
-        order_id: Production order to complete
-        user_email: User completing the order
-        quantity_good: Good quantity produced
-        quantity_scrapped: Scrapped quantity
-        force_close_short: Allow closing order with less than ordered quantity
-        notes: Completion notes
-
-    Returns:
-        Updated ProductionOrder
-    """
-    from app.services.inventory_service import process_production_completion
-
-    order = get_production_order(db, order_id)
-
-    # Idempotent: already complete is a no-op
-    if order.status == "complete":
-        return order
-
-    if order.status == "short":
-        raise HTTPException(
-            status_code=400,
-            detail="Order is in short status. Use the accept-short action to complete it."
-        )
-    if order.status not in ["in_progress", "released"]:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Cannot complete order in {order.status} status"
-        )
-
-    total_reported = quantity_good + quantity_scrapped
-    if total_reported > order.quantity_ordered:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Total reported ({total_reported}) exceeds ordered ({order.quantity_ordered})"
-        )
-
-    # Check for short completion
-    remaining = order.quantity_ordered - (order.quantity_completed or 0)
-    if quantity_good < remaining and not force_close_short:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Completing short ({quantity_good} of {remaining} remaining). Set force_close_short=true to close short."
-        )
-
-    # Process inventory transaction
-    try:
-        process_production_completion(
-            db=db,
-            production_order=order,
-            quantity_completed=Decimal(str(quantity_good)),
-            created_by=user_email,
-        )
-    except Exception as e:
-        logger.error(f"Failed to process production completion: {e}")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to process inventory: {str(e)}"
-        )
-
-    # Update order quantities
-    order.quantity_completed = (order.quantity_completed or 0) + quantity_good
-    order.quantity_scrapped = (order.quantity_scrapped or 0) + quantity_scrapped
-
-    # Check if complete (either fully or force-closed short)
-    if order.quantity_completed >= order.quantity_ordered or force_close_short:
-        order.status = "complete"
-        order.completed_at = datetime.now(timezone.utc)
-        order.actual_end = datetime.now(timezone.utc)
-
-        # Complete all operations
-        for op in order.operations:
-            if op.status != "complete":
-                op.status = "complete"
-                if not op.actual_end:
-                    op.actual_end = datetime.now(timezone.utc)
-
-        # Recalculate actual costs from consumed quantities and actual times
-        try:
-            from app.services.cost_estimation_service import recalculate_actual_cost
-            recalculate_actual_cost(db, order)
-        except Exception as e:
-            logger.warning("Actual cost recalculation failed for %s: %s", order.code, e)
-
-    if notes:
-        if order.notes:
-            order.notes = f"{order.notes}\n[{datetime.now(timezone.utc).isoformat()}] {notes}"
-        else:
-            order.notes = f"[{datetime.now(timezone.utc).isoformat()}] {notes}"
-
-    return order
-
-
-def accept_short_production_order(
-    db: Session,
-    order_id: int,
-    user_email: str,
-    user_id: int,
-    notes: Optional[str] = None,
-) -> ProductionOrder:
-    """Accept a production order short — complete it with the quantity already produced.
-
-    When all operations finish but quantity_completed < quantity_ordered, the PO
-    enters "short" status. No inventory transactions have happened yet at that point.
-    Accept-short processes inventory for the actual completed quantity and sets
-    the PO to "complete", unblocking downstream SO close-short.
-
-    Inventory actions:
-    - Releases all material reservations (allocated_quantity freed)
-    - Consumes materials for quantity_completed (BOM-proportional, may create GL via consume path)
-    - Receipts quantity_completed as finished goods (not quantity_ordered)
-    """
-    from app.services.inventory_service import (
-        consume_production_materials,
-        get_or_create_default_location,
-        create_inventory_transaction,
-        get_effective_cost_per_inventory_unit,
-    )
-    from app.models.close_short_record import CloseShortRecord
-
-    # Lock the row to prevent concurrent accept-short from double-applying inventory
-    order = (
-        db.query(ProductionOrder)
-        .filter(ProductionOrder.id == order_id)
-        .with_for_update()
-        .first()
-    )
-    if not order:
-        raise HTTPException(status_code=404, detail="Production order not found")
-
-    # Guard: must be in "short" status (all operations finished, qty < ordered)
-    if order.status != "short":
-        raise HTTPException(
-            status_code=400,
-            detail=f"Cannot accept short on order in '{order.status}' status. "
-                   f"Order must be in 'short' status."
-        )
-
-    # Guard: must have produced something but less than ordered
-    qty_completed = Decimal(str(order.quantity_completed or 0))
-    qty_ordered = Decimal(str(order.quantity_ordered))
-    if qty_completed <= 0:
-        raise HTTPException(
-            status_code=400,
-            detail="Cannot accept short: no units have been completed."
-        )
-    if qty_completed >= qty_ordered:
-        raise HTTPException(
-            status_code=400,
-            detail="Order is already fully completed — use the complete action instead."
-        )
-
-    # NOTE: Between accepting short on a component PO and its parent assembly PO,
-    # component available_quantity may be temporarily negative. This is expected —
-    # the assembly PO still holds reservations based on original ordered qty.
-    # The negative window resolves when the assembly PO is accepted-short.
-    # If this becomes a product feature, consider batch accept-short that resolves
-    # the full PO chain in a single transaction.
-
-    # Capture pre-action inventory state for audit record (all locations)
-    product_invs = db.query(Inventory).filter(
-        Inventory.product_id == order.product_id
-    ).all()
-    inv_snapshot = [
-        {
-            "product_id": order.product_id,
-            "location_id": inv.location_id,
-            "on_hand": str(inv.on_hand_quantity or 0),
-            "allocated": str(inv.allocated_quantity or 0),
-            "available": str(inv.available_quantity or 0),
-        }
-        for inv in product_invs
-    ]
-
-    # Idempotency guard: reject if inventory was already processed for this PO
-    from app.models.inventory import InventoryTransaction
-    existing_receipt = db.query(InventoryTransaction).filter(
-        InventoryTransaction.reference_type == "production_order",
-        InventoryTransaction.reference_id == order.id,
-        InventoryTransaction.transaction_type == "receipt",
-    ).first()
-    if existing_receipt:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Inventory already processed for {order.code}. Cannot accept short again."
-        )
-
-    # Process inventory for the actual completed quantity:
-    # 1. Release reservations and consume materials for qty_completed
-    consume_production_materials(
-        db=db,
-        production_order=order,
-        quantity_completed=qty_completed,
-        created_by=user_email,
-        release_reservations=True,
-    )
-
-    # 2. Receipt finished goods for qty_completed (NOT quantity_ordered)
-    #    We call create_inventory_transaction directly instead of receive_finished_goods
-    #    because receive_finished_goods always receipts quantity_ordered.
-    product = db.query(Product).filter(Product.id == order.product_id).first()
-    if not product:
-        raise HTTPException(
-            status_code=500,
-            detail=f"Product {order.product_id} not found for production order {order.code}"
-        )
-    location = get_or_create_default_location(db)
-    create_inventory_transaction(
-        db=db,
-        product_id=order.product_id,
-        location_id=location.id,
-        transaction_type="receipt",
-        quantity=qty_completed,
-        reference_type="production_order",
-        reference_id=order.id,
-        notes=f"Accept short PO#{order.code}: {qty_completed} of {qty_ordered} produced",
-        cost_per_unit=get_effective_cost_per_inventory_unit(product),
-        created_by=user_email,
-    )
-
-    # 3. Transition to complete
-    order.status = "complete"
-    order.completed_at = datetime.now(timezone.utc)
-    order.actual_end = datetime.now(timezone.utc)
-
-    # Complete any remaining operations
-    for op in order.operations:
-        if op.status != "complete":
-            op.status = "complete"
-            if not op.actual_end:
-                op.actual_end = datetime.now(timezone.utc)
-
-    # Append notes
-    if notes:
-        timestamp = datetime.now(timezone.utc).isoformat()
-        if order.notes:
-            order.notes = f"{order.notes}\n[{timestamp}] Accepted short: {notes}"
-        else:
-            order.notes = f"[{timestamp}] Accepted short: {notes}"
-
-    # Write audit record
-    audit_record = CloseShortRecord(
-        entity_type="production_order",
-        entity_id=order_id,
-        performed_by=user_id,
-        reason=notes,
-        line_adjustments=[{
-            "before_qty": str(qty_ordered),
-            "after_qty": str(qty_completed),
-            "reason": f"Accepted short: {qty_completed} of {qty_ordered} produced",
-        }],
-        inventory_snapshot=inv_snapshot,
-    )
-    db.add(audit_record)
-
-    return order
-
 
 def cancel_production_order(
     db: Session,
@@ -1257,97 +691,6 @@ def record_qc_inspection(
     logger.info(f"QC inspection recorded for {order.code}: {qc_status}")
 
     return inspection_result
-
-
-# =============================================================================
-# Split Order
-# =============================================================================
-
-def split_production_order(
-    db: Session,
-    order_id: int,
-    split_quantity: int,
-    user_email: str,
-    reason: Optional[str] = None,
-) -> tuple[ProductionOrder, ProductionOrder]:
-    """
-    Split a production order into two orders.
-
-    Args:
-        order_id: Order to split
-        split_quantity: Quantity for the new order
-        user_email: User performing the split
-        reason: Reason for split
-
-    Returns:
-        Tuple of (original_order, new_order)
-    """
-    from app.services.inventory_service import reserve_production_materials
-
-    order = get_production_order(db, order_id)
-
-    if order.status not in ["draft", "scheduled", "released"]:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Cannot split order in {order.status} status"
-        )
-
-    if split_quantity <= 0:
-        raise HTTPException(status_code=400, detail="Split quantity must be positive")
-
-    remaining = order.quantity_ordered - split_quantity
-    if remaining <= 0:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Split quantity ({split_quantity}) must be less than ordered ({order.quantity_ordered})"
-        )
-
-    # Update original order quantity
-    order.quantity_ordered = remaining
-
-    # Create new order
-    new_code = generate_production_order_code(db)
-    new_order = ProductionOrder(
-        code=new_code,
-        product_id=order.product_id,
-        bom_id=order.bom_id,
-        routing_id=order.routing_id,
-        sales_order_id=order.sales_order_id,
-        sales_order_line_id=order.sales_order_line_id,
-        quantity_ordered=split_quantity,
-        quantity_completed=0,
-        quantity_scrapped=0,
-        source="split",
-        status="draft",
-        priority=order.priority,
-        due_date=order.due_date,
-        assigned_to=order.assigned_to,
-        notes=f"Split from {order.code}" + (f": {reason}" if reason else ""),
-        created_by=user_email,
-    )
-    db.add(new_order)
-    db.flush()
-
-    # Copy operations with recalculated quantities
-    if order.routing_id:
-        copy_routing_to_operations(db, new_order, order.routing_id)
-
-    # Allocate materials for new order
-    reserve_production_materials(
-        db=db,
-        production_order=new_order,
-        created_by=user_email,
-    )
-
-    # Update original order notes
-    if order.notes:
-        order.notes = f"{order.notes}\n[SPLIT] {split_quantity} units moved to {new_code}"
-    else:
-        order.notes = f"[SPLIT] {split_quantity} units moved to {new_code}"
-
-    logger.info(f"Split {order.code}: {split_quantity} units to {new_code}")
-
-    return order, new_order
 
 
 # =============================================================================
@@ -2053,3 +1396,27 @@ def get_order_spools(db: Session, order_id: int) -> list[dict]:
             })
 
     return result
+
+
+# =============================================================================
+# Backward-compat re-exports (DEBT-1 D2-A mechanical split)
+# =============================================================================
+# release_production_order + its release-only self-heal helpers moved to
+# production_order_release_service.py; the start/complete/accept-short/split
+# lifecycle transitions moved to production_order_execution_service.py.
+# These imports keep the full public surface importable from this module so
+# existing callers (endpoints, sales_order_*, dispatch_service, inventory_service,
+# tests) continue to work unchanged. Placed at module end so this module finishes
+# defining its own names before the submodules import them back.
+from app.services.production_order_release_service import (  # noqa: E402,F401
+    _has_active_reservation,
+    _maybe_backfill_op_material_allocation,
+    _maybe_reserve_missing_materials,
+    release_production_order,
+)
+from app.services.production_order_execution_service import (  # noqa: E402,F401
+    start_production_order,
+    complete_production_order,
+    accept_short_production_order,
+    split_production_order,
+)
